requirements.txt
Flask==2.2.3
requests==2.31.0
psycopg2-binary==2.9.6
prettytable==3.6.0